requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
cryptography==38.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.2
django-appconf==1.0.5
django-bootstrap-modal-forms==2.2.0
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-cryptography==1.1
django-import-export==3.0.1
django-js-asset==2.0.0
django-outbox-menu==1.0.8
django-outbox-region==1.0.2
django-parler==2.3
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
outbox-encryption==1.0.7
Pillow==9.2.0
polib==1.1.1
pycparser==2.21
python-decouple==3.6
pytz==2022.5
PyYAML==6.0
shortuuid==1.0.9
six==1.16.0
sqlparse==0.4.3
tablib==3.2.1
urllib3==1.26.12
xlrd==2.0.1
xlwt==1.3.0
